Add Tacos Scaffold for Relationship Change Testing
When you have an existing has_many relationship with data in the db
and you switch to has_one, Rails will assign the first record as the
has_one. Unless you change data, all the records remain in the db
and will be restored if changed back to has_many.